+++ lab4/Lab4Task1Var2/Lab4Task1Var2/Circle.h
@@ -9,14 +9,14 @@ class CCircle final : public ISolidShape
 public:
 	CCircle(const V2f &centerCircle, float &radius, std::string const & lineColor, std::string const & fillColor);
 
-	double GetShapeArea() const override;
-	double GetShapePerimeter() const override;
+	float GetShapeArea() const override;
+	float GetShapePerimeter() const override;
 	std::string ToString() const override;
 	std::string GetLineColor() const override;
 	std::string GetInnerRegion() const override;
 private:
 	V2f m_centerCircle;
-	double m_radius;
+	float m_radius;
 	std::string m_lineColor;
 	std::string m_fillColor;
 };
